requirements.txt
gradio==4.44.0
pypdf==4.2.0
sentence-transformers==2.6.1
faiss-cpu==1.7.4
nltk==3.8.1
numpy==1.26.4
requests==2.31.0
torch
